backend/requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.2

# ─── Authentication & Security ────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# ─── Event Streaming ──────────────────────────────────────────────────────────
aiokafka==0.10.0
aio-pika==9.4.1            # RabbitMQ async

# ─── Computer Vision ──────────────────────────────────────────────────────────
ultralytics==8.2.18         # YOLOv8
opencv-python-headless==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.2.0

# ─── Firebase (FCM) ───────────────────────────────────────────────────────────
firebase-admin==6.5.0

# ─── LLM Integration ──────────────────────────────────────────────────────────
google-generativeai==0.5.4

# ─── Cloudinary (photo storage) ───────────────────────────────────────────────
cloudinary==1.40.0

# ─── Facial Recognition ───────────────────────────────────────────────────────
facenet-pytorch==2.6.0
torch==2.2.2
torchvision==0.17.2

# ─── Validation & Serialization ───────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── HTTP / WebSocket ─────────────────────────────────────────────────────────
httpx==0.27.0
websockets==12.0

# ─── Utilities ────────────────────────────────────────────────────────────────
structlog==24.1.0
tenacity==8.3.0
python-slugify==8.0.4

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
